+++ src/modules/category/category.service.ts
@@ -1,6 +1,8 @@
 import { Category } from "./entities/category.entity"
-import { NotFoundException } from "../../core/exceptions/base"
+import { NotFoundException, ConflictException } from "../../core/exceptions/base"
 import { EntityManager } from "typeorm"
+import { AppDataSource } from "../../config/database"
+import { SubCategory } from "../sub-category/entities/sub-category.entity"
 import { ICategoryRepository } from "./interfaces/category.repository.interface"
 
 export class CategoryService {
@@ -39,6 +41,10 @@ export class CategoryService {
 
     async delete(id: number): Promise<void> {
         await this.getById(id)
+        const subCategoryCount = await AppDataSource.getRepository(SubCategory).count({ where: { categoryId: id } })
+        if (subCategoryCount > 0) {
+            throw new ConflictException(`Cannot delete category, ${subCategoryCount} sub category(s) are still linked to this category`)
+        }
         await this.repository.delete(id)
     }
