+++ src/modules/sub-category/sub-category.service.ts
@@ -1,6 +1,8 @@
 import { SubCategory } from "./entities/sub-category.entity"
-import { NotFoundException } from "../../core/exceptions/base"
+import { NotFoundException, ConflictException } from "../../core/exceptions/base"
 import { EntityManager } from "typeorm"
+import { AppDataSource } from "../../config/database"
+import { Asset } from "../asset/entities/asset.entity"
 import { ISubCategoryRepository } from "./interfaces/sub-category.repository.interface"
 
 export class SubCategoryService {
@@ -39,6 +41,10 @@ export class SubCategoryService {
 
     async delete(id: number): Promise<void> {
         await this.getById(id)
+        const assetCount = await AppDataSource.getRepository(Asset).count({ where: { subCategoryId: id } })
+        if (assetCount > 0) {
+            throw new ConflictException(`Cannot delete sub category, ${assetCount} asset(s) are still linked to this sub category`)
+        }
         await this.repository.delete(id)
     }
